prepare for use in production env
- add start.sh shell script to start the app on the server
- add basic app conf variable to config.go
- take work directory as a cli argument
- cgi files
- add idea run config to the project
- add github runner workflow
- split cli and web versions
