+++ .env.dist
@@ -1,0 +1,5 @@
+ENVID=local
+APP_NAME=decision-maker
+APP_VERSION=dev
+HOSTNAME=localhost
+LOGGER_LEVEL=debug
